pyproject.toml
[project]
name = "rev-tun"
version = "0.1.0"
description = "Tool to register ssh reverse tunnels to supervisor"
authors = [{ name = 'observerw' }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepmerge>=2.0",
    "jinja2>=3.1.5",
    "pydantic>=2.10.5",
    "rich>=13.9.4",
    "tomli>=2.2.1",
    "typer-slim>=0.15.1",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
rev-tun = "rev_tun.cli:app"
